+++ CHANGELOG.md
@@ -8,6 +8,7 @@ _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Allowing `follow_by_tags` to interact with the user
 - Context manager to interaction calls in `like_by_tags` and `follow_likers`
+- Smart Hashtags based on locations `set_smart_location_hashtags`
 
 ### Changed
 - Expose `threaded_session` of Instapy.end()